refactor(types): split snapshot.ts into entities/system-snapshot.ts
